worked on the viewModel and my register fragment xml
