+++ app/src/main/java/com/example/tapshopping/core/ShopApplication.kt
@@ -5,5 +5,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class ShopApplication: Application() {
+class
+ShopApplication: Application() {
 }
