+++ app/src/main/java/com/example/tapshopping/ui/viewModel/UserViewModel.kt
@@ -1,10 +1,11 @@
-package com.example.tapshopping.ui.viewmodel
+package com.example.tapshopping.ui.viewModel
 
 import android.app.Application
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.tapshopping.data.model.DataModel
+import com.example.tapshopping.data.model.UserLoginData
 import com.example.tapshopping.data.model.UserRegistrationData
 import com.example.tapshopping.data.model.UsersResponse
 import com.example.tapshopping.domain.ShoppingRepository
@@ -14,15 +15,35 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class RegisterViewModel @Inject constructor(
+class UserViewModel @Inject constructor(
     application: Application, private val shoppingRepository: ShoppingRepository
-) : BaseViewModel(application) {
+) : ViewModel() {
 
     private val _user = MutableLiveData<Resource<UsersResponse>>()
     val users: LiveData<Resource<UsersResponse>> = _user
+    fun fetchUsers(userLogin: UserLoginData) {
+        viewModelScope.launch {
+            _user.value = Resource.loading()
+            val result = shoppingRepository.getUser(userLogin)
+            result.collect { values ->
+                if (values.isSuccess()) {
+
+                    _user.value = Resource.success(values.data)
+                    Resource.loading(true)
+                    Resource.error(" Unable to login user to the server", false)
+
+                } else {
+                    _user.value = Resource.error(message = values.message)
+                    Resource.loading(true)
+                    Resource.error(" Unable to login user to the server", false)
+                }
+            }
+        }
+    }
+
 
     fun registerUser(userRegistration: UserRegistrationData) {
-       Resource.loading(true)
+        Resource.loading(true)
         viewModelScope.launch {
             _user.value = Resource.loading()
             val result = shoppingRepository.createUser(userRegistration)
@@ -39,4 +60,5 @@ class RegisterViewModel @Inject constructor(
         }
     }
 
-    }
+
+}
