--- app/src/main/java/com/example/tapshopping/ui/viewmodel/BaseViewModel.kt
@@ -1,20 +1,0 @@
-package com.example.tapshopping.ui.viewmodel
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlin.coroutines.CoroutineContext
-
-open class BaseViewModel(application: Application): AndroidViewModel(application),
-    CoroutineScope {
-    private val job= Job()
-    override val coroutineContext: CoroutineContext
-        get() = job+ Dispatchers.Main
-
-    override fun onCleared() {
-        super.onCleared()
-        job.cancel()
-    }
-}
