+++ app/src/main/res/navigation/account.xml
@@ -37,6 +37,9 @@
         <action
             android:id="@+id/action_homeFragment_to_loginFragment"
             app:destination="@id/loginFragment" />
+        <action
+            android:id="@+id/action_accountFragment_to_logindUser"
+            app:destination="@id/loginUser" />
     </fragment>
 
     <fragment
@@ -53,4 +56,25 @@
         android:name="com.example.tapshopping.ui.fragment.CreateAdminFragment"
         android:label="fragment_create_admin"
         tools:layout="@layout/fragment_create_admin" />
+    <fragment
+        android:id="@+id/loginUser"
+        android:name="com.example.tapshopping.ui.fragment.LoginFragment"
+        android:label="fragment_login"
+        tools:layout="@layout/fragment_login" >
+        <action
+            android:id="@+id/action_logindUser_to_accountFragment"
+            app:destination="@id/accountFragment" />
+        <action
+            android:id="@+id/action_loginUser_to_registerUserFragment"
+            app:destination="@id/registerUserFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/registerUserFragment"
+        android:name="com.example.tapshopping.ui.fragment.RegisterFragment"
+        android:label="fragment_register_user"
+        tools:layout="@layout/fragment_register" >
+        <action
+            android:id="@+id/action_registerUserFragment_to_loginUser"
+            app:destination="@id/loginUser" />
+    </fragment>
 </navigation>
